# docs/API.yaml
openapi: 3.1.0
info:
  title: Wyvern Drive API
  description: |
    Internal API documentation for Wyvern Drive Supabase Edge Function.

    **⚠️ INTERNAL USE ONLY - NOT FOR PUBLIC DISTRIBUTION**

    This API handles file metadata storage, versioning, sharing, and video streaming.
    Actual file data is stored on Discord CDN; this API manages the index.
  version: 1.0.0
  contact:
    name: Developer
servers:
  - url: https://{project-ref}.supabase.co/functions/v1/api
    description: Production (Supabase Edge Function)
    variables:
      project-ref:
        default: your-project-ref
...
paths:
...
  /stream/{userId}/{fileId}:
    get:
      tags: [Streaming]
      summary: Stream video/audio with range requests
      description: |
        Supports HTTP Range requests for seeking in video/audio.
        Cannot stream encrypted files.
      operationId: streamFile
      parameters:
        - $ref: '#/components/parameters/userId'
        - $ref: '#/components/parameters/fileId'
        - name: Range
          in: header
          schema:
            type: string
            example: bytes=0-1023
      responses:
        '200':
          description: Full file content
          headers:
            Accept-Ranges:
              schema:
                type: string
                example: bytes
        '206':
          description: Partial content
          headers:
            Content-Range:
              schema:
                type: string
                example: bytes 0-1023/10240
        '400':
          description: Cannot stream encrypted files
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User identifier (hashed webhook URL)
      schema:
        type: string
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: integer
...